:truck: (modules mocks) move mocks to dedicated folders
