+++ frontend/src/modules/poi/mocks/index.ts
