+++ frontend/src/modules/poiType/mocks/index.ts
